+++ {{ cookiecutter.project_name }}/composition_repository/project_utils/urls.py
@@ -1,0 +1,18 @@
+from django.urls import path, include
+
+from project_composer.marker import EnabledApplicationMarker
+
+
+class ProjectUtilsUrls(EnabledApplicationMarker):
+    """
+    Project utils urls
+    """
+    def load_urlpatterns(self, urlpatterns):
+        """
+        Mount application urls
+        """
+        urlpatterns = super().load_urlpatterns(urlpatterns)
+
+        return [
+            path("utils/", include("project_utils.urls")),
+        ] + urlpatterns