+++ {{ cookiecutter.project_name }}/django-apps/project_utils/urls.py
@@ -1,0 +1,11 @@
+from django.urls import path
+
+from .views import ExposedSiteMetasView
+
+
+app_name = "project_utils"
+
+
+urlpatterns = [
+    path("site-metas/", ExposedSiteMetasView.as_view(), name="site-metas"),
+]