+++ {{ cookiecutter.project_name }}/django-apps/project_utils/views/__init__.py
@@ -1,0 +1,6 @@
+from .site_metas import ExposedSiteMetasView
+
+
+__all__ = [
+    "ExposedSiteMetasView",
+]